% Generated by roxygen2 (4.0.0): do not edit by hand
\name{integrate_times}
\alias{integrate_times}
\title{Adaptively Integrate an ODE System}
\usage{
integrate_times(stepper, ode_system, y, times, dt)
}
\arguments{
\item{stepper}{A \code{stepper} object, created by
\code{\link{stepper}} (other types will be supported
soon).}

\item{ode_system}{The ode_system system, created by
\code{\link{ode_system}}}

\item{y}{Initial conditions}

\item{times}{Vector of times.  First time is start time, last time
is end time.}

\item{dt}{Initial step size (will be tuned for controlled steppers
-- see odeint documentation)}
}
\description{
Integrate a system of ODEs at fixed times (perhaps adaptively).
This version \emph{always} saves state -- that is the reason for
calling it.
}
\author{
Rich FitzJohn
}

