% Generated by roxygen2 (4.0.0): do not edit by hand
\name{integrate_const}
\alias{integrate_adaptive}
\alias{integrate_const}
\alias{integrate_n_steps}
\alias{integrate_times}
\title{Integrate a System of ODEs}
\usage{
integrate_const(stepper, ode_system, y, t0, t1, dt, save_state = FALSE)

integrate_n_steps(stepper, ode_system, y, t0, dt, n, save_state = FALSE)

integrate_adaptive(stepper, ode_system, y, t0, t1, dt, save_state = FALSE)

integrate_times(stepper, ode_system, y, times, dt)
}
\arguments{
\item{stepper}{A \code{stepper} object, created by
\code{\link{make_stepper}}}

\item{ode_system}{The ode_system system, created by
\code{\link{ode_system}}}

\item{y}{Initial conditions}

\item{t0}{Time to start the integration}

\item{t1}{Time to finish the integration}

\item{dt}{Step size}

\item{n}{Number of steps to take (\code{integrate_n_steps} only)}

\item{times}{Vector of times (\code{integrate_times} only)}

\item{save_state}{Logical: return information about intermediate
points as an attribute.  Not applicable for
\code{integrate_times}.}
}
\description{
Integrate a system of ODEs
}
\details{
integrate_const: "Equidistant observer calls"

integrate_n_steps: "Integrate a given number of steps"

integrate_adaptive "Observer calls at each step"

integrate_times: "Observer calls at given time points"
}
\author{
Rich FitzJohn
}

