% Generated by roxygen2 (4.0.0): do not edit by hand
\name{stepper_algorithms}
\alias{stepper_algorithms}
\alias{stepper_basic_algorithms}
\alias{stepper_categories}
\alias{stepper_controlled_algorithms}
\alias{stepper_dense_algorithms}
\title{Lists of Stepper Algorithms}
\usage{
stepper_algorithms(category, have_jacobian = FALSE)

stepper_basic_algorithms(have_jacobian = FALSE)

stepper_controlled_algorithms(have_jacobian = FALSE)

stepper_dense_algorithms(have_jacobian = FALSE)

stepper_categories()
}
\arguments{
\item{category}{Either "basic" or "controlled"}

\item{have_jacobian}{Logical indicating if the stepper will work
on problems with a Jacobian (this is a superset: every problem
that provides a Jacobian will work fine with steppers that don't
require a Jacobian).}
}
\description{
Valid values for making a \code{\link{stepper}}.
}
\author{
Rich FitzJohn
}

