\docType{class}
\name{stepper_controlled-class}
\alias{make_stepper}
\alias{make_stepper_basic}
\alias{make_stepper_controlled}
\alias{stepper_controlled}
\alias{stepper_controlled-class}
\title{Controlled Stepper}
\usage{
make_stepper_basic(type)

make_stepper_controlled(type, atol = 1e-06, rtol = 1e-06)

make_stepper(category, type, ...)
}
\description{
Controlled stepper (documentation coming)
}

