\name{stepper_types}
\alias{stepper_basic_types}
\alias{stepper_controlled_types}
\alias{stepper_types}
\title{Types of Controlled Stepper}
\usage{
stepper_types(category)

stepper_controlled_types()

stepper_basic_types()
}
\arguments{
  \item{category}{Either "basic" or "controlled"}
}
\description{
Valid values for making a \code{\link{stepper_controlled}}.
}
\author{
Rich FitzJohn
}

