% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{stepper_deSolve-class}
\alias{make_stepper_deSolve}
\alias{make_stepper_deSolve_basic}
\alias{make_stepper_deSolve_controlled}
\alias{stepper_deSolve}
\alias{stepper_deSolve-class}
\title{Stepper (deSolve wrappers)}
\usage{
make_stepper_deSolve(category, algorithm, ...)

make_stepper_deSolve_basic(algorithm)

make_stepper_deSolve_controlled(algorithm, abs_tol = 1e-06, rel_tol = 1e-06)
}
\arguments{
\item{category}{Either "basic" or "controlled"}

\item{algorithm}{The stepper algorithm (e.g. "runge_kutta4")}

\item{...}{Additional parameters (still under development)}

\item{abs_tol}{Absolute tolerance (see atol in deSolve docs for now)}

\item{rel_tol}{Relative tolerance (see rtol in deSolve docs for now)}
}
\description{
Stepper (documentation coming)
}

