% Generated by roxygen2 (4.0.0): do not edit by hand
\name{stepper_types}
\alias{stepper_basic_types}
\alias{stepper_categories}
\alias{stepper_controlled_types}
\alias{stepper_types}
\title{Types of Controlled Stepper}
\usage{
stepper_types(category)

stepper_controlled_types()

stepper_basic_types()

stepper_categories()
}
\arguments{
\item{category}{Either "basic" or "controlled"}
}
\description{
Valid values for making a \code{\link{stepper}}.
}
\author{
Rich FitzJohn
}

